shop

+++ controllers/shopControllers.js
@@ -14,6 +14,9 @@ const shopRegister = async (req, res, next) => {
                 await User.findByIdAndUpdate(req.user?._id, {
                     $push: { adminShop: { $each: [created._id], $position: 0 } }
                 }, { new: true });
+                await User.findByIdAndUpdate(req.user?._id, {
+                    $unset: { sellerShop: "" }
+                }, { new: true });
                 return res.status(200).json({ message: `Shop Registration Successfully! Your '${name}' shop is approved!`, data: created });
             }
         }
@@ -32,6 +35,9 @@ const shopRegister = async (req, res, next) => {
                     const user = await User.findOne({ _id: req.user._id });
                     user.sellerShop = created._id;
                     await user.save();
+                    await User.findByIdAndUpdate(req.user?._id, {
+                        $unset: { adminShop: "" },
+                    }, { new: true });
                     return res.status(200).json({ message: `Shop Registration Successfully! Your '${name}' shop is Under Review. You will receive Confirmation Soon.`, data: created });
                 }
             }
@@ -46,17 +52,21 @@ const shopRegister = async (req, res, next) => {
 const updateShop = async (req, res, next) => {
     const { name, phone, address, openDate, closeDate, email } = req.body;
     try {
-        const shop = await Shop.findOne({ user: req.user._id });
-        if (!shop) return res.status(404).json({ error: { "shop": "shop not founds!" } })
-        if (shop) {
-            const shopUpdated = await Shop.findByIdAndUpdate(req.params.id, {
-                name, phone, address, openDate, closeDate, email
-            }, { new: true });
-            if (!shopUpdated) {
-                return res.status(400).json({ error: { "shop": "shop updated failed!" } })
-            }
-            if (shopUpdated) {
-                return res.status(200).json({ message: "shop updated successfully!", data: shopUpdated })
+        if (!(req?.user?.role === 'seller' || req?.user?.isAdmin === true)) {
+            return res.status(400).json({ error: { "shop": "Permission denied! Buyers do not update the store." } })
+        } else {
+            const shop = await Shop.findOne({ user: req.user._id });
+            if (!shop) return res.status(404).json({ error: { "shop": "shop not founds!" } })
+            if (shop) {
+                const shopUpdated = await Shop.findByIdAndUpdate(req.params.id, {
+                    name, phone, address, openDate, closeDate, email
+                }, { new: true });
+                if (!shopUpdated) {
+                    return res.status(400).json({ error: { "shop": "shop not founds!" } })
+                }
+                if (shopUpdated) {
+                    return res.status(200).json({ message: "shop updated successfully!", data: shopUpdated })
+                }
             }
         }
     } catch (error) {
@@ -65,24 +75,31 @@ const updateShop = async (req, res, next) => {
 }
 const shopRemove = async (req, res, next) => {
     try {
-        if (req?.user?.isAdmin === true) {
-            const removeIt = await Shop.findByIdAndRemove(req.params.id);
-            if (!removeIt) return res.status(400).json({ error: { "shop": "Shop not founds!" } });
-            if (removeIt) {
-                await User.findByIdAndUpdate(req.user?._id, {
-                    $pull: { adminShop: { $each: [req.params.id], $position: 0 } }
-                }, { new: true });
-                return res.status(200).json({ message: `Shop successfully removed!` });
+        if (!(req?.user?.role === 'seller' || req?.user?.isAdmin === true)) {
+            return res.status(400).json({ error: { "shop": "Permission denied! Buyers do not remove the store." } })
+        } else {
+            if (req?.user?.isAdmin === true) {
+                const removeIt = await Shop.findByIdAndRemove(req.params.id);
+                if (!removeIt) return res.status(400).json({ error: { "shop": "Shop not founds!" } });
+                if (removeIt) {
+                    await User.findByIdAndUpdate(req.user?._id, {
+                        $pull: { adminShop: req.params.id },
+                    }, { new: true });
+                    return res.status(200).json({ message: `Shop successfully removed!` });
+                }
             }
-        }
-        if (req?.user?.role === 'seller') {
-            const removeIt = await Shop.findByIdAndRemove(req.params.id);
-            if (!removeIt) return res.status(400).json({ error: { "shop": "Shop not founds!" } });
-            if (removeIt) {
-                await User.findByIdAndUpdate(req.user?._id, {
-                    $unset: { sellerShop: "" }
-                }, { new: true });
-                return res.status(200).json({ message: `Shop successfully removed!` });
+            if (req?.user?.role === 'seller') {
+                const removeIt = await Shop.findByIdAndRemove(req.params.id);
+                if (!removeIt) return res.status(400).json({ error: { "shop": "Shop not founds!" } });
+                if (removeIt) {
+                    await User.findByIdAndUpdate(req.user?._id, {
+                        $unset: { sellerShop: "" }
+                    }, { new: true });
+                    await User.findByIdAndUpdate(req.user?._id, {
+                        $unset: { adminShop: "" },
+                    }, { new: true });
+                    return res.status(200).json({ message: `Shop successfully removed!` });
+                }
             }
         }
     } catch (error) {
